+++ src/components/text-editor/__internal__/__ui__/Placeholder/placeholder.style.ts
@@ -1,6 +1,7 @@
 import styled from "styled-components";
 
 const StyledPlaceholder = styled.div`
+  pointer-events: none;
   position: absolute;
   top: 12px;
   color: var(--colorsUtilityYin055);