+++ src/locales/locale.ts
@@ -136,9 +136,11 @@ interface Locale {
     undo: () => string;
   };
   tabs: {
-    error: () => string;
-    warning: () => string;
-    info: () => string;
+    error: (tabTitle?: string) => string;
+    warning: (tabTitle?: string) => string;
+    info: (tabTitle?: string) => string;
+    scrollLeftText?: () => string;
+    scrollRightText?: () => string;
   };
   textEditor: {
     boldAria: () => string;
@@ -190,10 +192,16 @@ interface Locale {
     };
   };
   sort: {
-    accessibleName: (
+    /**
+     * @deprecated The `sort.accessibleName` locale string is deprecated in favour of using `aria-live` regions.
+     * Use an `aria-live` region to announce changes in sort order to assistive technologies.
+     */
+    accessibleName?: (
       sortContent?: string,
       sortType?: "ascending" | "descending",
     ) => string;
+    /* TODO make ariaRoleDescription required as part of FE-7599 */
+    ariaRoleDescription?: () => string;
   };
   splitButton: {
     ariaLabel: () => string;